refactor(perf): split src/session/manager.ts — extract manager-deps.ts and manager-run.ts
